# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AFRICA CDC KnowledgeHub API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/knowledge_hub",
            "description": "Current deployment (set APP_URL in .env)"
        }
    ],
    "paths": {
...
        "/api/experts": {
            "get": {
                "tags": [
                    "Experts"
                ],
                "summary": "List Country Experts",
                "description": "Returns a list of all Country Experts",
                "operationId": "ListCountryExperts",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "description": "Search term for specific records",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Experts"
                ],
                "summary": "Create Country Expert",
                "description": "Allows users to submit country experts for admin approval",
                "operationId": "CreateExpert",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "job_title": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "country_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request, when some required data is missing"
                    },
                    "404": {
                        "description": "Not found when you sent the request to an invalid endpoint"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/experts/{id}": {
            "get": {
                "tags": [
                    "Experts"
                ],
                "summary": "Show Expert",
                "description": "Retrieve a single expert record based on the ID.",
                "operationId": "ShowExpert",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expert ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Experts"
                ],
                "summary": "Update Country Expert",
                "description": "Allows users to update Country Expert Record",
                "operationId": "UpdateExpert",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Record Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "job_title": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    },
                                    "country_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request, when some required data is missing"
                    },
                    "404": {
                        "description": "Not found when you sent the request to an invalid endpoint"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}